+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and releases in PushmiPullyu adheres to [Semantic Versioning](https://semver.org
 
 - Changed Danger token in Github Actions
 - remove pg dependency
+- unblock CI (rubocop and coveralls)
 
 ## [2.0.2] - 2021-11-22